New tests

+++ 1.LabOOP/Lab1_3/Program.cs
@@ -19,8 +19,8 @@ namespace Lab1_3
             Console.WriteLine($"Double answer: {ansDouble}");
 
             //Получаем float результаты
-            float ansFloat = (float)((Pow(a - b, 4) - (Pow(a, 4) + 6 * Pow(a, 2) * Pow(b, 2) + Pow(b, 4))) /
-                (-4 * a * Pow(b, 3) - 4 * Pow(a, 3) * b));
+            float ansFloat = (((float)Pow(a - b, 4) - (float)(Pow(a, 4) + (float)6 * (float)Pow(a, 2) * (float)Pow(b, 2) + (float)Pow(b, 4))) /
+                (-4 * (float)a * (float)Pow(b, 3) - 4 * (float)Pow(a, 3) * (float)b));
             Console.WriteLine($"Float answer: {ansFloat}");
         }
     }
